Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrack/language-expected.txt

CONSOLE MESSAGE: The language contains a null character and is not a valid BCP 47 language tag.
CONSOLE MESSAGE: The language contains a null character and is not a valid BCP 47 language tag.

PASS TextTrack.language
PASS TextTrack.language, \u0000
